+++ crates/melody_compiler/src/ast/consts.rs
@@ -1,1 +1,2 @@
 pub static NOT: &str = "not";
+pub static LAZY: &str = "lazy";