+++ crates/melody_compiler/src/regex.rs
@@ -1,0 +1,3 @@
+pub mod ast_to_regex;
+mod utils;
+pub use self::ast_to_regex::to_regex;