+++ crates/melody_compiler/src/source.rs
@@ -1,3 +1,0 @@
-pub mod ast_to_source;
-mod utils;
-pub use self::ast_to_source::to_source;